make pointsdisplay folder and add styled file
